move auth middleware to pkg
