+++ components/director/internal/domain/operation/converter_test.go
@@ -2,11 +2,12 @@ package operation_test
 
 import (
 	"database/sql"
+	"testing"
+
 	"github.com/kyma-incubator/compass/components/director/internal/domain/operation"
 	"github.com/kyma-incubator/compass/components/director/internal/model"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"testing"
 )
 
 func TestEntityConverter_ToEntity(t *testing.T) {
