+++ components/director/internal/domain/operation/repository.go
@@ -2,12 +2,13 @@ package operation
 
 import (
 	"context"
+	"time"
+
 	"github.com/kyma-incubator/compass/components/director/internal/model"
 	"github.com/kyma-incubator/compass/components/director/internal/repo"
 	"github.com/kyma-incubator/compass/components/director/pkg/apperrors"
 	"github.com/kyma-incubator/compass/components/director/pkg/log"
 	"github.com/kyma-incubator/compass/components/director/pkg/resource"
-	"time"
 )
 
 const operationTable = `public.operation`
