+++ components/director/internal/operations_manager/opcreator.go
@@ -3,11 +3,12 @@ package operationsmanager
 import (
 	"context"
 	"encoding/json"
+	"time"
+
 	"github.com/kyma-incubator/compass/components/director/internal/model"
 	"github.com/kyma-incubator/compass/components/director/pkg/log"
 	"github.com/kyma-incubator/compass/components/director/pkg/persistence"
 	"github.com/pkg/errors"
-	"time"
 )
 
 const (
