+++ components/director/internal/operations_manager/opdata.go
@@ -2,6 +2,7 @@ package operationsmanager
 
 import (
 	"encoding/json"
+
 	"github.com/pkg/errors"
 )
