+++ components/director/internal/systemfetcher/template.go
@@ -5,6 +5,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+
 	"k8s.io/client-go/util/jsonpath"
 
 	"github.com/imdario/mergo"
